Drop deprecated defaults setters and getters from Stoplight::Light
